12/6
This week, we reread the PID tutorial and discussed a new condition for the while loop controlling the PID algorithm: when the error is 0, wait a few milliseconds, if the error is still 0, end the loop.

Next week: Start testing to find kP constant, then use given equations to find the rest of the constants.